Finish pinning nightly Rust version, fix new clippy warnings (#56)
* Cleanup nightly clippy warnings

* Fix the rest of the nightly warnings

* Bump back nightly version, b/c binary is 1kb smaller somehow

+++ crates/smb2_practice_mod/src/mods/stage_edits.rs
@@ -117,7 +117,7 @@ impl StageEdits {
                 if (*mkb::stagedef).goal_count < 1 {
                     return;
                 }
-                self.rev_goal_idx %= (*mkb::stagedef).goal_count as u32;
+                self.rev_goal_idx %= (*mkb::stagedef).goal_count;
 
                 let start = (*mkb::stagedef).start;
                 let goal = (*mkb::stagedef).goal_list.add(self.rev_goal_idx as usize);
